pyproject.toml
[project]
name = "portfolio-backend"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "django-filter>=25.1",
    "django-stubs-ext>=5.1.3",
    "django-unfold>=0.53.0",
    "django-valkey>=0.2.1",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "psycopg[binary,pool]>=3.2.6",
    "toml>=0.10.2",
]

[dependency-groups]
dev = [
    "django-stubs>=5.1.3",
    "djangorestframework-stubs[compatible-mypy]>=3.15.3",
    "faker>=37.1.0",
    "mypy>=1.15.0",
    "mypy-extensions>=1.0.0",
    "ruff>=0.11.2",
    "types-toml>=0.10.8.20240310",
]

[tool.ruff]
lint.select = ["E", "F", "I", "C", "B", "DJ"]
line-length = 120
show-fixes = true
cache-dir = "~/.cache/ruff"
exclude = [".venv"]

[tool.mypy]
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = [
    ".venv/",
    "migrations/",
    "tests/",
    "wsgi.py",
    "manage.py",
    "db/",
]

plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
module = ["unfold.*"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "core.settings"
